tic-tac-toe update December 25 2017 8:39 PM
- Developed and tested fully functional game API with
required calls to the server (curl and web app). A user can now create a
new game, retrieve a list of completed and incomplete games, and
specify a game to view the outcome of/continue playing.
- Next step is ensuring that a visual display of game statistics
including #of games won per player/number of draws is achieved for
each user.
- Final steps include tackling multiplayer mode and other bonus
requirements before moving on to CSS/ReadMe/Documentation.