tic-tac-toe update January 4 2018 2:51 PM
- Developed multiplayer functionality to include a gameWatcher variable
that calls a resourceWatcher function and listens for a change to
the update of a game board in another player's browser (player o).
Changes made are updated dynamically to the game board of the player
creating the game (player x). However, as of now, the gameWatcher
connection to the API is not closing in a timely manner after game
completion, thus resulting in internal server errors (500). (When
I get time, I need to take a closer look into closing gameWatcher
connections to the API)
- Developed bootstrap modal popup to display (and enable retrieval) of
completed and incomplete games on button click.
- Improved navigation functionality to various sections of the game
page.
- Enhanced CSS to include box-shadow effects for the game table and
results table. Included borders around respective sections.
- Next step is to work on documentation and validation of text box
inputs.
- Also need to ensure that the game is multi device.